// frontend/src/api/generated/models/outbound-slip-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { OutboundSlipStatus } from './outbound-slip-status';
// May contain unused imports in some cases
// @ts-ignore
import type { OutboundSlipType } from './outbound-slip-type';

/**
 * 
 * @export
 * @interface OutboundSlipSummary
 */
export interface OutboundSlipSummary {
    /**
     * 出荷伝票ID
     * @type {number}
     * @memberof OutboundSlipSummary
     */
    'id'?: number;
    /**
     * 伝票番号
     * @type {string}
     * @memberof OutboundSlipSummary
     */
    'slipNumber'?: string;
    /**
     * 
     * @type {OutboundSlipType}
     * @memberof OutboundSlipSummary
     */
    'slipType'?: OutboundSlipType;
    /**
     * 出荷先取引先名（振替の場合はnull）
     * @type {string}
     * @memberof OutboundSlipSummary
     */
    'partnerName'?: string | null;
    /**
     * 出荷予定日
     * @type {string}
     * @memberof OutboundSlipSummary
     */
    'plannedDate'?: string;
    /**
     * 
     * @type {OutboundSlipStatus}
     * @memberof OutboundSlipSummary
     */
    'status'?: OutboundSlipStatus;
    /**
     * 明細件数
     * @type {number}
     * @memberof OutboundSlipSummary
     */
    'lineCount'?: number;
    /**
     * 作成日時
     * @type {string}
     * @memberof OutboundSlipSummary
     */
    'createdAt'?: string;
}