// frontend/src/api/generated/models/outbound-slip-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { OutboundSlipSummary } from './outbound-slip-summary';

/**
 * 
 * @export
 * @interface OutboundSlipPageResponse
 */
export interface OutboundSlipPageResponse {
    /**
     * 
     * @type {Array<OutboundSlipSummary>}
     * @memberof OutboundSlipPageResponse
     */
    'content'?: Array<OutboundSlipSummary>;
    /**
     * 
     * @type {number}
     * @memberof OutboundSlipPageResponse
     */
    'page'?: number;
    /**
     * 
     * @type {number}
     * @memberof OutboundSlipPageResponse
     */
    'size'?: number;
    /**
     * 
     * @type {number}
     * @memberof OutboundSlipPageResponse
     */
    'totalElements'?: number;
    /**
     * 
     * @type {number}
     * @memberof OutboundSlipPageResponse
     */
    'totalPages'?: number;
}